+++ x13/cuda_x13_fugue512_alexis.cu
@@ -245,11 +245,11 @@ static void SMIX_LDG(const uint32_t shared[4][256], uint32_t &x0,uint32_t &x1,ui
 __global__ __launch_bounds__(256,3)
 void x13_fugue512_gpu_hash_64_alexis(int *thr_id, uint32_t threads, uint64_t *g_hash)
 {
-	if ((*(int*)(((uint64_t)thr_id) & ~15ULL)) & (1 << (((uint64_t)thr_id) & 15)))
+	if ((*(int*)(((uintptr_t)thr_id) & ~15ULL)) & (1 << (((uintptr_t)thr_id) & 15)))
 		return;
 	__shared__ uint32_t shared[4][256];
 
-	//	if(threadIdx.x<256){
+//	if(threadIdx.x<256){
 		const uint32_t tmp = mixtab0[threadIdx.x];
 		shared[0][threadIdx.x] = tmp;
 		shared[1][threadIdx.x] = ROR8(tmp);
@@ -405,18 +405,18 @@ void x13_fugue512_gpu_hash_64_final_alexis(uint32_t threads,const uint32_t* __re
 				resNonce[1] = tmp;		
 		}
 	}
-} 
+}
 
 __host__
-void x13_fugue512_cpu_hash_64_alexis(int *thr_id, uint32_t threads, uint32_t *d_hash){
-
+void x13_fugue512_cpu_hash_64_alexis(int *thr_id, uint32_t threads, uint32_t *d_hash)
+{
 	const uint32_t threadsperblock = 256;
 
 	// berechne wie viele Thread Blocks wir brauchen
 	dim3 grid((threads + threadsperblock-1)/threadsperblock);
 	dim3 block(threadsperblock);
 
-	x13_fugue512_gpu_hash_64_alexis << <grid, block >> >(thr_id, threads, (uint64_t*)d_hash);
+	x13_fugue512_gpu_hash_64_alexis<<<grid, block>>>(thr_id, threads, (uint64_t*)d_hash);
 }
 
 __host__
