+++ x15/cuda_x14_shabal512_alexis.cu
@@ -105,7 +105,7 @@ void ROTATE(uint32_t* A){
 __global__ __launch_bounds__(384,3)
 void x14_shabal512_gpu_hash_64_alexis(int *thr_id, uint32_t threads, uint32_t *g_hash)
 {
-	if ((*(int*)(((uint64_t)thr_id) & ~15ULL)) & (1 << (((uint64_t)thr_id) & 15)))
+	if ((*(int*)(((uintptr_t)thr_id) & ~15ULL)) & (1 << (((uintptr_t)thr_id) & 15)))
 		return;
 	const uint32_t thread = (blockDim.x * blockIdx.x + threadIdx.x);
 
@@ -177,7 +177,7 @@ __host__ void x14_shabal512_cpu_hash_64_alexis(int *thr_id, uint32_t threads, ui
 	dim3 grid((threads + threadsperblock-1)/threadsperblock);
 	dim3 block(threadsperblock);
 
-	x14_shabal512_gpu_hash_64_alexis << <grid, block >> >(thr_id, threads, d_hash);
+	x14_shabal512_gpu_hash_64_alexis<<<grid, block>>>(thr_id, threads, d_hash);
 }
 
 __global__ __launch_bounds__(512,2)
